@@ -167,8 +167,8 @@ func (s Bytes) Take(n int) (out Bytes) {
 
 // ----------------- CONVERTING METHOD -----------------
 
-// S convert s into []interface{}
-func (s Bytes) S() (out []interface{}) {
+// S convert s into []any
+func (s Bytes) S() (out []any) {
 	for _, v := range s {
 		out = append(out, v)
 	}
